+++ spec/dummy/config/routes.rb
@@ -1,3 +1,5 @@
 Rails.application.routes.draw do
   mount Unmagic::Passkeys::Engine => "/unmagic-passkeys"
+
+  use_unmagic_passkeys
 end